tools/requirements.txt
OpenCC==1.1.7
pypinyin==0.50.0
regex==2023.12.25
pandas
python-dotenv
requests
